:sparkles: ADD: SuaRevenda purchase add/get

+++ C#/SuaRevenda/Models/Purchase.cs
@@ -2,7 +2,8 @@ namespace SuaRevenda.Models
 {
     public class Purchase : Origin
     {
-        public long Price { get; set; }
+        public double Price { get; set; }
+
         public override void makeSell()
         {
             throw new NotImplementedException();
